pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "obsidian-vault-pipeline"
version = "0.4.0"
description = "全自动Obsidian知识管理Pipeline - 生产级知识管理流水线"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "OpenClaw", email = "openclaw@example.com" },
]
keywords = [
    "obsidian",
    "knowledge-management",
    "pipeline",
    "llm",
    "markdown",
    "zettelkasten",
    "evergreen",
    "moc",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business :: Groupware",
    "Topic :: Text Processing :: Markup",
    "Topic :: Utilities",
]

dependencies = [
    "anthropic>=0.21.0",
    "openai>=1.0.0",
    "litellm>=1.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.28.0",
    "pyyaml>=6.0",
    "feedparser>=6.0.0",
    "beautifulsoup4>=4.12.0",
    "tiktoken>=0.5.0",
    "watchdog>=3.0.0",
    "networkx>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/fakechris/obsidian_vault_pipeline"
Documentation = "https://github.com/fakechris/obsidian_vault_pipeline#readme"
Repository = "https://github.com/fakechris/obsidian_vault_pipeline"
Issues = "https://github.com/fakechris/obsidian_vault_pipeline/issues"

[project.scripts]
ovp = "openclaw_pipeline.unified_pipeline_enhanced:main"
ovp-pipeline = "openclaw_pipeline.unified_pipeline_enhanced:main"
ovp-article = "openclaw_pipeline.auto_article_processor:main"
ovp-github = "openclaw_pipeline.auto_github_processor:main"
ovp-paper = "openclaw_pipeline.auto_paper_processor:main"
ovp-evergreen = "openclaw_pipeline.auto_evergreen_extractor:main"
ovp-moc = "openclaw_pipeline.auto_moc_updater:main"
ovp-quality = "openclaw_pipeline.batch_quality_checker:main"
ovp-clippings = "openclaw_pipeline.clippings_processor:main"
ovp-autopilot = "openclaw_pipeline.autopilot.daemon:main"
ovp-lint = "openclaw_pipeline.lint_checker:main"
ovp-query = "openclaw_pipeline.query_tool:main"
ovp-repair = "openclaw_pipeline.commands.repair:main"
ovp-query-to-wiki = "openclaw_pipeline.query_to_wiki:main"
ovp-batch-evergreen = "openclaw_pipeline.batch_evergreen:main"
ovp-rebuild-registry = "openclaw_pipeline.commands.rebuild_registry:main"
ovp-migrate-links = "openclaw_pipeline.commands.migrate_broken_links:main"
ovp-merge-duplicates = "openclaw_pipeline.commands.merge_duplicates:main"
ovp-promote-candidates = "openclaw_pipeline.promote_candidates:main"
ovp-graph = "openclaw_pipeline.graph_cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/openclaw_pipeline"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
